+++ src/components/RestaurantAddModal/RestaurantAddModal.css
@@ -47,92 +47,3 @@
 .modal-title {
   margin-bottom: 24px;
 }
-
-.form-item {
-  display: flex;
-  flex-direction: column;
-
-  margin-bottom: 24px;
-}
-
-.form-item label {
-  color: var(--grey-400);
-  font-size: 14px;
-}
-
-.form-item--required label::after {
-  padding-left: 4px;
-
-  color: var(--primary-color);
-  content: "*";
-}
-
-.form-item .help-text {
-  color: var(--grey-300);
-}
-
-.form-item input,
-.form-item textarea,
-.form-item select {
-  padding: 8px;
-  margin: 6px 0;
-
-  border: 1px solid var(--grey-200);
-  border-radius: 8px;
-
-  font-size: 16px;
-}
-
-.form-item textarea {
-  resize: none;
-}
-
-.form-item select {
-  height: 44px;
-
-  padding: 8px;
-
-  border: 1px solid var(--grey-200);
-  border-radius: 8px;
-
-  color: var(--grey-300);
-}
-
-input[name="name"],
-input[name="link"] {
-  height: 44px;
-}
-
-.button-container {
-  display: flex;
-}
-
-.button {
-  width: 100%;
-  height: 44px;
-
-  margin-right: 16px;
-
-  border: none;
-  border-radius: 8px;
-
-  font-weight: 600;
-  cursor: pointer;
-}
-
-.button:last-child {
-  margin-right: 0;
-}
-
-.button--secondary {
-  border: 1px solid var(--grey-300);
-  background: transparent;
-
-  color: var(--grey-300);
-}
-
-.button--primary {
-  background: var(--primary-color);
-
-  color: var(--grey-100);
-}
