+++ src/constants/eventType.ts
@@ -2,4 +2,5 @@ export const CUSTOM_EVENT_TYPE = {
   sortChange: "sortChange",
   addRestaurant: "addRestaurant",
   filterCategory: "filterCategory",
+  resetForm: "resetForm",
 } as const;
