+++ src/constants/regexp.ts
@@ -1,0 +1,1 @@
+export const OPTION_ELEMENT_REGEXP = /<option value=(\d+)>(\d+)<\/option>/g;
